backend-rest-django/requirements.txt
astroid==2.0.4
django>=2.1.5
django-environ==0.4.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
logilab-astng==0.24.3
logilab-common==1.4.2
mccabe==0.6.1
mysqlclient==1.3.13
pep8==1.7.1
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
six==1.11.0
wrapt==1.10.11
